Reorganized persistence e2e test and tweaked health check to cover for worfklows that end very quickly, like callbackstatetimeouts
